requirements.txt
anthropic
python-dotenv
pandas
numpy
tqdm
openai
langchain
qdrant-client
sentence-transformers
scipy
matplotlib
seaborn
pyyaml
streamlit